+++ CHANGELOG.md
@@ -35,5 +35,6 @@
 - Distribute with .dtx instead of .sty
 - Reordered build, install, distribute, and uninstall sections of the README
 - Pulled build up one level of headings in the README
+- fix some typos in README